+++ TestsPlugin/ManusReference.swift
@@ -1,0 +1,1 @@
+../Tests/CodexBarTests/ManusReference.swift