+++ TestsPlugin/T3ChatReference.swift
@@ -1,0 +1,1 @@
+../Tests/CodexBarTests/T3ChatReference.swift